refactor: separate TYPES constant
