+++ src/lib/types.ts
@@ -1,0 +1,15 @@
+export const TYPES = {
+  IMemberAdapter: 'IMemberAdapter',
+  INotificationAdapter: 'INotificationAdapter',
+  IAccountAdapter: 'IAccountAdapter',
+  IFinancialSourceAdapter: 'IFinancialSourceAdapter',
+  IChartOfAccountAdapter: 'IChartOfAccountAdapter',
+  IFinancialTransactionHeaderAdapter: 'IFinancialTransactionHeaderAdapter',
+  IMemberRepository: 'IMemberRepository',
+  INotificationRepository: 'INotificationRepository',
+  IAccountRepository: 'IAccountRepository',
+  IFinancialSourceRepository: 'IFinancialSourceRepository',
+  IChartOfAccountRepository: 'IChartOfAccountRepository',
+  IFinancialTransactionHeaderRepository: 'IFinancialTransactionHeaderRepository',
+  AuditService: 'AuditService'
+};
